Extract compact output helpers partial
