+++ src/CodeIndex/Cli/QueryCommandRunner.cs
@@ -352,28 +352,6 @@ public static partial class QueryCommandRunner
         OutputFormatJson,
         OutputFormatCompact,
     };
-    private static int GetCompactSectionLimit(QueryCommandOptions options)
-        => options.LimitExplicit ? options.Limit : DefaultCompactSectionLimit;
-
-    private static int GetCompactSourceLimit(int compactLimit)
-    {
-        var sourceLimit = compactLimit + 1;
-        return NumericFlagUpperBounds.TryGetValue("--limit", out var maxLimit)
-            ? Math.Min(sourceLimit, maxLimit)
-            : sourceLimit;
-    }
-
-    private static JsonObject ApplySymbolAnalysisCompactCaps(SymbolAnalysisResult analysis, int sectionLimit)
-    {
-        var sections = new JsonObject();
-        TruncateCompactSection(analysis.Definitions, sectionLimit, sections, "definitions");
-        TruncateCompactSection(analysis.NearbySymbols, sectionLimit, sections, "nearby_symbols");
-        TruncateCompactSection(analysis.References, sectionLimit, sections, "references");
-        TruncateCompactSection(analysis.Callers, sectionLimit, sections, "callers");
-        TruncateCompactSection(analysis.Callees, sectionLimit, sections, "callees");
-        return BuildCompactTruncationMetadata(sectionLimit, sections);
-    }
-
     private static JsonObject ApplyOutlineCompactCaps(OutlineResult outline, int sectionLimit)
         => ApplyOutlineSymbolLimit(outline, sectionLimit);
 
@@ -733,20 +711,6 @@ public static partial class QueryCommandRunner
         return payload;
     }
 
-    private static JsonObject BuildCompactTruncationMetadata(int sectionLimit, JsonObject sections)
-        => new()
-        {
-            ["section_limit"] = sectionLimit,
-            ["sections"] = sections,
-        };
-
-    private static void AddCompactJsonFields(JsonObject payload, int compactLimit, JsonObject truncation)
-    {
-        payload["compact"] = true;
-        payload["compact_limit"] = compactLimit;
-        payload["truncation"] = truncation;
-    }
-
     private static void AddJsonByteLimitField(JsonObject payload, QueryCommandOptions options)
     {
         if (options.MaxJsonBytes.HasValue)
@@ -804,20 +768,6 @@ public static partial class QueryCommandRunner
             AddReplayValueOption(args, "--min-entrypoint-confidence", options.MinEntrypointConfidence.ToString("0.###", CultureInfo.InvariantCulture));
         if (options.MaxJsonBytes.HasValue)
             AddReplayValueOption(args, "--max-json-bytes", options.MaxJsonBytes.Value.ToString(CultureInfo.InvariantCulture));
-    }
-
-    private static void TruncateCompactSection<T>(List<T> items, int sectionLimit, JsonObject sections, string sectionName)
-    {
-        var sourceCount = items.Count;
-        if (sourceCount > sectionLimit)
-            items.RemoveRange(sectionLimit, sourceCount - sectionLimit);
-
-        sections[sectionName] = new JsonObject
-        {
-            ["returned"] = items.Count,
-            ["source_count"] = sourceCount,
-            ["truncated"] = sourceCount > sectionLimit,
-        };
     }
 
     private static void ApplyInspectFieldSelection(JsonObject payload, QueryCommandOptions options, JsonSerializerOptions jsonOptions)
